refactor: type 선언 추가
- 여러 곳에서 선언되던 enum, class 등을 한 곳에서 선언하는 것으로 수정
