+++ server/src/dto/order.dto.ts
@@ -1,32 +1,5 @@
-enum SIZE {
-  small = 's',
-  medium = 'm',
-  large = 'l',
-}
-
-enum TEMPERATURE {
-  hot = 'h',
-  ice = 'c',
-}
-
-enum PAYMENT {
-  card = 'card',
-  cash = 'cash',
-}
-
-interface option {
-  size: SIZE;
-  temperature: TEMPERATURE;
-}
-
-interface food {
-  id: number;
-  name: string;
-  unit: number;
-  options: option;
-  eachPrice: number;
-}
+import { FOOD, PAYMENT } from 'src/types';
 
 export class OrderDto {
-  constructor(readonly foods: food[], readonly payment: PAYMENT, readonly date: Date) {}
+  constructor(readonly foods: FOOD[], readonly payment: PAYMENT, readonly date: Date) {}
 }
